Adds nowayout support to PCF8563W driver.
The nowayout mechanism allows a user to configure the RTC such that
it can't be shut down after it is started. This can be useful for
certain usecases such as ensuring the recomputer can recover if
a UPS shutdown is triggered and power is restored before the
UPS discharges sufficiently to cut power.